refactor: tidy accrual util, gate policies UI, split assign-modal form
- Move accrual-preview types (AccrualPreviewRow/AccrualEvent/ScheduleConfig/
  CalendarUnit) to LeavePolicyTypes and its constants (ACCRUAL_PREVIEW_ROW_LIMIT,
  CALENDAR_UNIT, INTERVAL_STEP) to leavePolicyConstants; the util now holds only
  functions.
- Reintroduce useIsLeavePoliciesEnabled as the single source of truth and gate
  UserLeavePolicies vs UserAssignedLeaveTypes on it (defaults enabled until the
  backend LEAVE_POLICY org-config flag exists).
- Extract the assign-modal body into AssignLeavePolicyForm, leaving the modal to
  own state, buttons and mutations.
- Drop the redundant .trim() on the composed employee name.
