+++ frontend/src/community/leave/components/molecules/IndividualEmployeeLeaveReportSection/IndividualEmployeeLeaveReportSection.tsx
@@ -3,9 +3,11 @@ import { FC, useEffect, useMemo, useState } from "react";
 import PeopleLayout from "~community/common/components/templates/PeopleLayout/PeopleLayout";
 import { useTranslator } from "~community/common/hooks/useTranslator";
 import { useGetLeaveTypes } from "~community/leave/api/LeaveApi";
+import UserAssignedLeaveTypes from "~community/leave/components/molecules/UserAssignedLeaveTypes/UserAssignedLeaveTypes";
 import UserLeaveHistory from "~community/leave/components/molecules/UserLeaveHistory/UserLeaveHistory";
 import UserLeavePolicies from "~community/leave/components/molecules/UserLeavePolicies/UserLeavePolicies";
 import UserLeaveUtilization from "~community/leave/components/molecules/UserLeaveUtilization/UserLeaveUtilization";
+import useIsLeavePoliciesEnabled from "~community/leave/hooks/useIsLeavePoliciesEnabled";
 import { useLeaveStore } from "~community/leave/store/store";
 import { LeaveType } from "~community/leave/types/CustomLeaveAllocationTypes";
 import UpgradeOverlay from "~enterprise/common/components/molecules/UpgradeOverlay/UpgradeOverlay";
@@ -33,6 +35,8 @@ const IndividualEmployeeLeaveReportSection: FC<Props> = ({
   );
 
   const { isAtLeastCoreTier } = useTier();
+
+  const isLeavePoliciesEnabled = useIsLeavePoliciesEnabled();
 
   const { resetLeaveRequestParams } = useLeaveStore((state) => state);
 
@@ -68,12 +72,16 @@ const IndividualEmployeeLeaveReportSection: FC<Props> = ({
         <>
           <h2 className="h2 text-black">{translateText(["pageHead"])}</h2>
 
-          <UserLeavePolicies
-            employeeId={selectedUser}
-            employeeName={`${employeeFirstName ?? ""} ${
-              employeeLastName ?? ""
-            }`.trim()}
-          />
+          {isLeavePoliciesEnabled ? (
+            <UserLeavePolicies
+              employeeId={selectedUser}
+              employeeName={`${employeeFirstName ?? ""} ${
+                employeeLastName ?? ""
+              }`}
+            />
+          ) : (
+            <UserAssignedLeaveTypes employeeId={selectedUser} pageSize={8} />
+          )}
 
           {leaveTypesList?.length > 0 && (
             <UserLeaveUtilization
